requirements.txt
greenlet==2.0.1
SQLAlchemy==1.4.45
python-dotenv==0.21.0
